GafferOSL : Extract CapturingErrorHandler from OSLExpressionEngine.
It's still private to GafferOSL, but can now be used in the internals of other classes.
